Modification du controller deleteBook

+++ controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/book');
+const fs = require('fs'); 
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book);
@@ -50,7 +51,18 @@ exports.modifyBook = (req, res, next) => {
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre supprimé avec succès !' }))
-        .catch((error) => res.status(400).json({ error }));
-}
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId !== req.auth.userId) {
+                res.status(401).json({ message: 'Non autorisé' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Livre supprimé avec succès !' }))
+                        .catch((error) => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch((error) => res.status(500).json({ error }));
+};
